Concurrency and MPI testing files
Added the version of code used to test concurrency. This program ran on both computer and raspberry pi

Added the version of code used for MPI testing. This program ran exclusively on the raspberry pi since getting MPI to work on windows was a nightmare, -especially given the short time we had.